[#7] Simple policy and way to change participation status

+++ app/controllers/events_controller.rb
@@ -1,16 +1,19 @@
 class EventsController < ApplicationController
   def index
     @events = Event.all
+    authorize Event
   end
   
   def show
     @event = Event.find(params[:id])
+    authorize @event
   end
   
   def new
     @event = Event.new
     @event.event_games.build
     @event.event_participation.build
+    authorize @event
     
     render 'form'
   end
@@ -27,12 +30,14 @@ class EventsController < ApplicationController
   
   def edit
     @event = Event.find(params[:id])
+    authorize @event
     
     render 'form'
   end
   
   def update
     @event = Event.find(params[:id])
+    authorize @event
     
     if @event.update_attributes(event_params)
       redirect_to events_path
@@ -43,6 +48,7 @@ class EventsController < ApplicationController
   
   def destroy
     @event = Event.find(params[:id])
+    authorize @event
     
     if @event.destroy
       redirect_to events_path
@@ -51,10 +57,27 @@ class EventsController < ApplicationController
     end
   end
   
+  def change_participation
+    @event_participation = EventParticipation.find(params[:id])
+    authorize @event_participation, :update?
+    
+    if @event_participation.update(event_participation_params)
+      flash.notice = 'Success'
+    else
+      flash.alert = 'Failure'
+    end
+
+    redirect_back(fallback_location: events_path)
+  end
+  
   private
   
   def event_params
     params.require(:event).permit! # TODO: Fix this
+  end
+  
+  def event_participation_params
+    params.require(:event_participation).permit! # TODO: Fix this
   end
   
   def arel
